int/utils/setops: removes unused.
