+++ internal/evolution/genetics/compatibles.go
@@ -40,9 +40,20 @@ func marktypes(fd *ast.FuncDecl) []types.NodeType {
 	return l
 }
 
+func Intersect[T comparable](l, r []T) []T {
+	ml := prepLookupMap(l)
+	i := make([]T, 0, min(len(l), len(r)))
+	for _, v := range r {
+		if _, found := ml[v]; found {
+			i = append(i, v)
+		}
+	}
+	return i
+}
+
 func mutualFieldTypes(fd1, fd2 *ast.FuncDecl) ([]types.NodeType, error) {
 	t1, t2 := marktypes(fd1), marktypes(fd2)
-	mutuals := setops.Intersect(t1, t2)
+	mutuals := Intersect(t1, t2)
 	if len(mutuals) == 0 {
 		return nil, fmt.Errorf("There is no single field type mutually existing in the entire subtree of two function declarations")
 	}
